+++ Sources/SwiftTitleCase/String+TitleCase+MLA.swift
@@ -8,57 +8,27 @@ import Foundation
 
 internal extension String {
     enum MLA: TitleCaseWordList, TitleCaseStyleType {
+        static let prepositionsRange: any RangeExpression<Int> = 1...
+
         static let options: String.TitleStyleOptions = [
-            .firstLastCapitalized,
-            .lowerCase3orLessCharacters
+            .firstAndLastCapitalized,
+            .lowerCase3orLessCharacters,
+            .useLessCommonPrepositions,
         ]
 
         internal static var wordListSet: Set<String>  = [
             "a",        // Article
-            "about",    // Preposition
-            "about",    // Preposition
-            "above",    // Preposition
-            "above",    // Preposition
-            "across",   // Preposition
-            "after",    // Preposition
-            "against",  // Preposition
-            "among",    // Preposition
             "an",       // Article
             "and",      // Conjunction
-            "around",   // Preposition
             "as",       // Conjunction
-            "at",       // Preposition
-            "before",   // Preposition
-            "behind",   // Preposition
-            "below",    // Preposition
-            "beside",   // Preposition
-            "between",  // Preposition
             "but",      // Conjunction
-            "by",       // Preposition
-            "down",     // Preposition
-            "during",   // Preposition
-            "for",      // Conjunction, Preposition
-            "from",     // Preposition
+            "for",      // Conjunction
             "if",       // Conjunction
-            "in",       // Preposition
-            "inside",   // Preposition
-            "into",     // Preposition
-            "near",     // Preposition
             "nor",      // Conjunction
-            "of",       // Preposition
-            "off",      // Preposition
-            "on",       // Preposition
             "or",       // Conjunction
             "out",      // Preposition
             "over",     // Preposition
-            "so",       // Conjunction
             "the",      // Article
-            "through",  // Preposition
-            "to",       // Preposition
-            "toward",   // Preposition
-            "under",    // Preposition
-            "up",       // Preposition
-            "with",     // Preposition
             "yet",      // Conjunction
         ]
     }